requirements.txt
langchain==0.1.0
langgraph==0.0.26
langchain-google-genai==0.0.11
langchain-community==0.0.13
wikipedia==1.4.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
nest-asyncio==1.5.8
mysql-connector-python==8.2.0